+++ scripts/lib/d3glycanstructure.js
@@ -1,11 +1,60 @@
 //Draws the structure
-function d3glycanstructure(glycanname, drawdivID) {
+function d3glycanstructure(glycanname, options) {
   //MOD: this version of d3glycanstructure is incapable of drawing glycopeptides
   // // aminoacids.some(function (v) { return glycanname.search(v) > -1 }) // check for aminoacids
   // if (glycanname.search(/(\{)|(\})/g) > -1) {
   //   gpdraw(glycanname, drawdivID);
   //   return;
   // }
+
+  //default configurations for drawing
+  let configuration = {
+    width: 360,
+    height: 520,
+    orientation: 'bottom-to-top',
+    drawdivID: 'd3glycanstruc',
+    symbsize: 30,
+    drawingareachoice: 'fixglycansize',
+    fucopt: 'fucdown',
+    linkageVisible: true,
+    linkRotate: false,
+    linkAbbr: false,
+    linkFontSize: 16,
+    margin: {
+      top: 50,
+      right: 50,
+      bottom: 50,
+      left: 50,
+    },
+    drawingSettingsDivID: 'drawsetting',
+    userOverrideOptions: true
+  }
+
+  //Put all of the options into a variable called configuration
+  if ('undefined' !== typeof options) {
+    for (var i in options) {
+      if ('undefined' !== typeof options[i]) { configuration[i] = options[i]; }
+    }//for i
+  }//if
+
+  if (configuration.userOverrideOptions === true) {
+    configuration = getUserDrawingOptions(configuration);
+  }
+
+  if (configuration.orientation === 'right-to-left') {
+    let height = configuration.width;
+    configuration.width = configuration.height;
+    configuration.height = height;
+  }
+
+
+  // console.log(configuration);
+
+
+  // if (configuration.orientation === "right-to-left") {
+  //   configuration.margin.right = configuration.margin.right + 20;
+  // }
+
 
   drawersize = 400;
   if (glyd3render === true) {
@@ -23,58 +72,86 @@ function d3glycanstructure(glycanname, drawdivID) {
   var depth = getDepth(data); //using the depth you can set different dimensions for the SVG images
   //allowing for better visibility of larger structures
 
-  var margin = {
-    top: 50,
-    right: 50,
-    bottom: 50,
-    left: 50,
-  },
-    width = 360 - margin.left - margin.right;
+  var margin = configuration.margin;
+
+  let height, width;
+  if (configuration.orientation == 'right-to-left') {
+    height = configuration.height - margin.top - margin.bottom;
+  } else {
+    width = configuration.width - margin.left - margin.right;
+  }
   //if fixed drawing area height:
-  var drawingareachoice = document.getElementById('drawingchoice').value;
-  if (drawingareachoice === 'fixdrawingarea') {
-    height = 520 - margin.top - margin.bottom;
-  } else if (drawingareachoice === 'fixglycansize' && depth <= 2) {
-    height = 50 * depth;
-  } else if (drawingareachoice === 'fixglycansize' && depth < 10) {
-    height = 50 * depth;
-  } else if (drawingareachoice === 'fixglycansize' && depth >= 10) {
-    height = 520 - margin.top - margin.bottom;
+  if (configuration.drawingareachoice === 'fixdrawingarea') {
+    if (configuration.orientation == 'right-to-left') {
+      width = configuration.width - margin.left - margin.right;
+    } else {
+      height = configuration.height - margin.top - margin.bottom;
+    }
+  } else if (configuration.drawingareachoice === 'fixglycansize' && depth <= 2) {
+    if (configuration.orientation == 'right-to-left') {
+      width = 50 * depth;
+    } else {
+      height = 50 * depth;
+    }
+  } else if (configuration.drawingareachoice === 'fixglycansize' && depth < 10) {
+    if (configuration.orientation == 'right-to-left') {
+      width = 50 * depth;
+    } else {
+      height = 50 * depth;
+    }
+  } else if (configuration.drawingareachoice === 'fixglycansize' && depth >= 10) {
+    if (configuration.orientation == 'right-to-left') {
+      width = configuration.width - margin.top - margin.bottom;
+    } else {
+      height = configuration.height - margin.top - margin.bottom;
+    }
+
   }
 
   var orientations = {
-
-    size: [width, height],
-    x: function (d) {
-      return d.x;
+    "right-to-left": {
+      size: [height, width],
+      x: function (d) { return width - d.y; },
+      y: function (d) { return d.x; }
     },
-    y: function (d) {
-      return height - d.y;
-    },
+    "bottom-to-top": {
+      size: [width, height],
+      x: function (d) {
+        return d.x;
+      },
+      y: function (d) {
+        return height - d.y;
+      },
+    }
 
   };//rotates the glycan to vertical drawing from bottom to top
 
   //only refresh for mouseover type of drawings not for the ctrl+click type
-  if (drawdivID == 'd3glycanstruc') {
-    d3.select('#d3glycanstrucsub' + drawdivID).remove();
+  if (configuration.drawdivID == 'd3glycanstruc') {
+    d3.select('#d3glycanstrucsub' + configuration.drawdivID).remove();
     //    d3.select('#' + drawdivID).select('.refreshremove').remove();
+  }
+
+  if (configuration.orientation === "bottom-to-top") {
+    d3.select(`#${configuration.drawdivID}`)
+      .style('flex-direction', 'column-reverse')
+  } else {
+    d3.select(`#${configuration.drawdivID}`)
+      .style('flex-direction', 'row-reverse')
   }
 
   //create a div to draw the glycans
   var div = d3
-    .select('#' + drawdivID)
+    .select('#' + configuration.drawdivID)
     .append('div')
-    .attr('id', 'd3glycanstrucsub' + drawdivID);
-  //push the div to the bottom if it is in the drawer
-  if (drawdivID == 'd3glycanstruc') {
-    div.style('position', 'absolute')
-      .style('bottom', '10px');
-  }
+    .attr('id', 'd3glycanstrucsub' + configuration.drawdivID);
+
+
 
   var svg = d3
-    .select('#d3glycanstrucsub' + drawdivID)
+    .select('#d3glycanstrucsub' + configuration.drawdivID)
     .append('svg')
-    .attr('id', drawdivID + 'SVG')
+    .attr('id', configuration.drawdivID + 'SVG')
     .attr('width', width + margin.left + margin.right)
     .attr('height', height + margin.top + margin.bottom)
     .append('g')
@@ -82,12 +159,8 @@ function d3glycanstructure(glycanname, drawdivID) {
     .attr('transform', 'translate(' + margin.left + ',' + margin.top + ')');
 
 
-  //adjust the symbol size depending on the depth
-  var symbsize = document.getElementById('symbolsize').value;
 
-
-  var o = orientations;
-
+  var o = orientations[configuration.orientation];
 
   // Compute the layout.
   var treemap = d3.tree().size(o.size);
@@ -95,16 +168,41 @@ function d3glycanstructure(glycanname, drawdivID) {
   var nodes = d3.hierarchy(data);
 
 
+  if (configuration.orientation == 'right-to-left') {
+    nodes = sortchildren(nodes, true);
+  }
+
+
   nodes = treemap(nodes);
 
-  fixfucers(nodes, width);
+  // oxford = true;
+  // if (oxford === true) {
+  //   nodes = getoxfordcoords(nodes, width, height, symbsize);
+  // }
+
+  fixfucers(nodes, configuration);
 
   //adjust to keep root node at the same position.
-  var dxg = nodes.x - width / 2;
-  if (dxg != 0) {
-    d3.select('#glycang')
-      .attr('transform', 'translate(' + (margin.left - dxg) + ',' + margin.top + ')');
+  if (configuration.orientation == 'right-to-left') {
+    var dxg = nodes.x - height / 2;
+    if (dxg != 0) {
+      if (depth < 10) {
+        d3.select('#glycang')
+          .attr('transform', 'translate(' + (width + margin.left - 50 * depth) + ',' + (margin.top - dxg) + ')');
+      } else {
+        d3.select('#glycang')
+          .attr('transform', 'translate(' + margin.left + ',' + (margin.top - dxg) + ')');
+      }
+    }
+  } else {
+    var dxg = nodes.x - width / 2;
+    if (dxg != 0) {
+      d3.select('#glycang')
+        .attr('transform', 'translate(' + (margin.left - dxg) + ',' + margin.top + ')');
+    }
   }
+
+
 
   var links = nodes.descendants().slice(1);
 
@@ -117,7 +215,7 @@ function d3glycanstructure(glycanname, drawdivID) {
     .attr('class', 'glyd3link')
     .attr('d', function (d) {
       return (
-        'M' + d.x + ',' + o.y(d) + 'L' + d.parent.x + ',' + o.y(d.parent)
+        'M' + o.x(d) + ',' + o.y(d) + 'L' + o.x(d.parent) + ',' + o.y(d.parent)
       );
     })
     .attr('stroke-width', 2);
@@ -129,17 +227,17 @@ function d3glycanstructure(glycanname, drawdivID) {
     .enter()
     .append('g')
     .attr('transform', function (d) {
-      return rotateFuc(d, o);
+      return rotateFuc(d, o, configuration);
     });
 
 
   node
     .append('use')
     .attr('xlink:href', function (d, i) { return drawsymbol(d, i, 'glycan') })
-    .attr('x', -symbsize / 2)
-    .attr('y', -symbsize / 2)
-    .attr('height', symbsize) //important for firefox
-    .attr('width', symbsize)
+    .attr('x', -configuration.symbsize / 2)
+    .attr('y', -configuration.symbsize / 2)
+    .attr('height', configuration.symbsize) //important for firefox
+    .attr('width', configuration.symbsize)
     .attr('class', 'symbol')
     .on('click', function (d) {
       var path = nodes.path(d);
@@ -164,7 +262,7 @@ function d3glycanstructure(glycanname, drawdivID) {
     .on('mouseover', function (d) {
       d3.select(this.parentElement).append('circle')
         .attr('class', 'highlightnode')
-        .attr('r', (symbsize/2)+10)
+        .attr('r', (configuration.symbsize / 2) + 10)
         .attr('fill', function (d) {
           var modelist = Array.from(document.querySelectorAll('#modelist input'));
           var mode = modelist.length && modelist.find(r => r.checked).value;
@@ -196,42 +294,64 @@ function d3glycanstructure(glycanname, drawdivID) {
       d3.selectAll('.highlightnode').remove();
     });
 
-  // //get the linkage drawing settings;
-  var linkVisible, linkRotate, linkAbbr, linkFontSize;
-  (document.getElementById('linkagevisible') != null) ? linkVisible = document.getElementById('linkagevisible').checked : linkVisible = true;
-  (document.getElementById('linkagerotate') != null) ? linkRotate = document.getElementById('linkagerotate').checked : linkRotate = true;
-  (document.getElementById('linkageabbr') != null) ? linkAbbr = document.getElementById('linkageabbr').checked : linkAbbr = false;
-  (document.getElementById('linkFontSize') != null) ? linkFontSize = document.getElementById('linkFontSize').value : linkFontSize = 14;
 
-
-  if (linkVisible === true) {
+  if (configuration.linkageVisible === true) {
     //append linkage text
     node
       .append('text')
       .attr('transform', function (d, i) {
-        return transformlinkText(d, i, linkRotate, linkFontSize, linkAbbr,symbsize);
+        return transformlinkText(d, i, configuration);
       })
-      .text(function (d) { 
-        // console.log(linkageText(d, linkAbbr));
-        return linkageText(d, linkAbbr) })
-      .attr('font-size', linkFontSize)
+      .text(function (d) {
+        return linkageText(d, i, configuration)
+      })
+      .attr('font-size', configuration.linkFontSize)
       .attr('text-anchor', 'middle')
       .style('font-family', 'Arial,Sans');
   }
 
   //append substitution text
-  node
+  let subtext = node
     .append('text')
     .text(function (d) { return subText(d); })
-    .attr('font-size', linkFontSize)
-    .style('font-family', 'Arial,Sans')
-    .attr('x', -linkFontSize - 3)
-    .attr('y', 3)
-    .attr('text-anchor', 'end');
+    .attr('font-size', configuration.linkFontSize)
+    .style('font-family', 'Arial,Sans');
+
+  if (configuration.orientation === 'bottom-to-top') {
+    subtext.attr('x', -configuration.linkFontSize - 3)
+      .attr('y', 3)
+      .attr('text-anchor', 'end');
+  } else {
+    subtext.attr('x', 3)
+      .attr('y', configuration.linkFontSize + configuration.symbsize / 2)
+      .attr('text-anchor', 'middle');
+  }
   //  console.log(glycanname);
   //  console.log('drawing finished');
 }
 
+function getUserDrawingOptions(options) {
+  let settingsForm = document.querySelector(`#${options.drawingSettingsDivID} form`);
+  if (!settingsForm) {
+    console.warn('GlycoGlyph: No settings div found, using default options.');
+    return options
+  };
+  let inputs = settingsForm.querySelectorAll('input, select, checkbox');
+
+  inputs.forEach(f => {
+    if (f.type === 'checkbox') {
+      options[f.id] = f.checked;
+    } else if (f.type === 'number') {
+      options[f.id] = +f.value;
+    } else {
+      options[f.id] = f.value;
+    }
+  })
+
+  //console.log(options);
+
+  return options;
+}
 
 
 //getDepth function gets how deep the data is nested
@@ -249,18 +369,20 @@ function getDepth(obj) {
 };
 
 //fix the fucoses to right angles
-function fixfucers(node, width) {
+function fixfucers(node, options) {
   //try these name 
   //Galb1-4(Galb1-4)GlcNAcb1-3(Fuca1-2)Sp8
   //Galb1-4(Fuca1-3)GlcNAcb1-6(Fuca1-4(Fuca1-2Galb1-3)GlcNAcb1-3)Galb1-4Glc-Sp21
   //Fuca1-4(Fuca1-2Galb1-3)GlcNAcb1-2Mana1-3(Fuca1-4(Fuca1-2Galb1-3)GlcNAcb1-2Mana1-3)Manb1-4GlcNAcb1-4GlcNAcb-Sp19
   //Fuca1-2Galb1-4(Fuca1-3)GlcNAcb1-3Galb1-4(Fuca1-3)GlcNAcb1-3Galb1-4(Fuca1-3)GlcNAcb-Sp0
 
-  //get the option for the fucose
-  var fucopt = document.getElementById('fucoseopt').value;
   //if original styling then just return and don't modify the values for the fucose
-  if (fucopt === 'fucoriginal') { return }
+  if (options.fucopt === 'fucoriginal') { return }
 
+  let width = options.width - options.margin.left - options.margin.right;
+  if (options.orientation === "right-to-left") {
+    width = options.height - options.margin.top - options.margin.bottom;
+  }
   //get the center value
   var centerx = node.x;
 
@@ -281,38 +403,38 @@ function fixfucers(node, width) {
       //get the parent node x value
       var parentx = d.parent.x;
 
-      if (fucopt === 'fucout') {
+      if (options.fucopt === 'fucout') {
         //this code puts the fucoses outwards at all times
         if (d.x < centerx) {
-          d.x = d.parent.x - (width / 4); //to the left 
+          d.x = d.parent.x - (width / 5); //to the left 
         } else if (d.x > centerx) {
-          d.x = d.parent.x + (width / 4); //to the right
+          d.x = d.parent.x + (width / 5); //to the right
         } else if (d.x == centerx) {
           if (d.parent.x < centerx) {
-            d.x = d.parent.x - (width / 4); //to the left 
+            d.x = d.parent.x - (width / 5); //to the left 
           } else if (d.parent.x > centerx) {
-            d.x = d.parent.x + (width / 4); //to the left 
+            d.x = d.parent.x + (width / 5); //to the left 
           } else {
-            d.x = d.parent.x - (width / 4); //to the left as default condition
+            d.x = d.parent.x - (width / 5); //to the left as default condition
           }
         } else {
-          d.x = d.parent.x - (width / 4); //to the left as default condition
+          d.x = d.parent.x - (width / 5); //to the left as default condition
         }
       }
 
-      if (fucopt === 'fucdown') {
+      if (options.fucopt === 'fucdown') {
         //This code can shift fucose either to left or right of parent
         //check if fucose is left or right of parent node and adjust the x-value with respect to the width
         if (d.x < d.parent.x) {
-          d.x = d.parent.x - (width / 4); //to the left
+          d.x = d.parent.x - (width / 5); //to the left
         } else {
-          d.x = d.parent.x + (width / 4); //to the right
+          d.x = d.parent.x + (width / 5); //to the right
         }
       }
 
-      if (fucopt === 'fucleft') {
+      if (options.fucopt === 'fucleft') {
         //This code is alternative to above to push fucose strictly to the left of the parent
-        d.x = d.parent.x - (width / 4); //push to left
+        d.x = d.parent.x - (width / 5); //push to left
       }
 
       //find the difference in x between the siblings of the fucose and the parent
@@ -336,14 +458,55 @@ function fixfucers(node, width) {
 }
 
 //returns coordinates for transform for linkage text
-function transformlinkText(d, i, linkRotate, fontSize, linkAbbr,symbsize) {
+function transformlinkText(d, i, options) {
   // console.log(d);
   if (i > 0) { //ignores the first sugar as that is the root
     //calculate the angle between this node and the parent node
     var angle = Math.atan2((d.y - d.parent.y), (d.parent.x - d.x)) * 180 / Math.PI;
+    // console.log(angle)
+    // console.log(d.data.name);
+
+    if (options.orientation === "right-to-left") {
+      //enter special logic to calculate the text angles for right to left type orientation
+
+      var n;
+      options.linkAbbr ? n = 2 : n = 4;
+
+      if (d.data.monosaccharide === "Fuc" && options.fuctype != 'fucoriginal') {
+        if (angle <= 45) {
+          angle = 90;
+          return `translate(${(d.y - d.parent.y) / 2 + options.linkFontSize / 2},${(d.parent.x - d.x) / 2}) rotate(${angle})`;
+        } else {
+          angle = -90;
+          return `translate(${(d.y - d.parent.y) / 2 - options.linkFontSize / 2},${(d.parent.x - d.x) / 2}) rotate(${angle})`;
+
+        }
+      }
+
+      if (options.linkRotate == false) {
+        switch (true) {
+          case (d.x < d.parent.x):
+            return `translate(${(d.y - d.parent.y) / 2 + options.linkFontSize},${(d.parent.x - d.x) / 2 - options.linkFontSize})`;
+          case (d.x === d.parent.x):
+            return `translate(${(d.y - d.parent.y) / 2},${(d.parent.x - d.x) / 2 - options.linkFontSize / 2})`;
+          case (d.x > d.parent.x):
+            return `translate(${(d.y - d.parent.y) / 2 + options.linkFontSize},${(d.parent.x - d.x) / 2 + options.linkFontSize})`;
+        }
+      }
+
+      angle = Math.atan2((d.parent.x - d.x), (d.y - d.parent.y)) * 180 / Math.PI;
+      // console.log(`new angle: ${angle}`)
+      // console.log(`dy = ${d.y} \t dparenty = ${d.parent.y} \t dx = ${d.x} \t dparentx = ${d.parent.x}\n
+      //   x = ${Math.abs(d.parent.x - d.x) / 2} \t , \t y = ${Math.abs(d.parent.y - d.y) / 2}
+      //   `);
+
+      return `translate(${(d.y - d.parent.y) / 2},${(d.parent.x - d.x) / 2 - options.linkFontSize / 2}) rotate(${angle})`;
+
+      // end right-to-left block
+    }
+
     //adjust angle for fucose if not using classical drawing
-    var fuctype = document.getElementById('fucoseopt').value;
-    if (d.data.monosaccharide === "Fuc" && fuctype != 'fucoriginal' && d.depth > 0) {
+    if (d.data.monosaccharide === "Fuc" && options.fuctype != 'fucoriginal' && d.depth > 0) {
 
       if (d.x < d.parent.x) {
         angle = 30 //to the left
@@ -351,7 +514,7 @@ function transformlinkText(d, i, linkRotate, fontSize, linkAbbr,symbsize) {
         angle = 150 //to the right
       }
     }
-    if (d.data.monosaccharide === "Fuc" && fuctype != 'fucoriginal') {
+    if (d.data.monosaccharide === "Fuc" && options.fuctype != 'fucoriginal') {
       if (angle <= 45) {
         return 'translate(' + (d.parent.x - d.x - 10) / 2 + ',' + (d.y - d.parent.y + 25) / 2 + ') rotate(' + angle + ')';
       } else {
@@ -359,33 +522,36 @@ function transformlinkText(d, i, linkRotate, fontSize, linkAbbr,symbsize) {
         return 'translate(' + (d.parent.x - d.x + 10) / 2 + ',' + (d.y - d.parent.y + 30) / 2 + ') rotate(' + angle + ')';
       }
     }
-    if (linkRotate == false) {
+    if (options.linkRotate == false) {
       var n;
-      linkAbbr ? n = 2 : n = 4;
+      options.linkAbbr ? n = 2 : n = 4;
       switch (true) {
         case (d.x < d.parent.x):
-          return 'translate(' + (d.parent.x - d.x - (fontSize * 6) + (fontSize * n)) / (fontSize / 2) + ',' + (d.y - d.parent.y + 5) / 2 + ')';
+          return 'translate(' + (d.parent.x - d.x - (options.linkFontSize * 6) + (options.linkFontSize * n)) / (options.linkFontSize / 2) + ',' + (d.y - d.parent.y + 5) / 2 + ')';
         case (d.x === d.parent.x):
-          return 'translate(' + (d.parent.x - d.x + (fontSize * 6) + (fontSize * n)) / (fontSize / 2) + ',' + (d.y - d.parent.y + 5) / 2 + ')';
+          return 'translate(' + (d.parent.x - d.x + (options.linkFontSize * 6) + (options.linkFontSize * n)) / (options.linkFontSize / 2) + ',' + (d.y - d.parent.y + 5) / 2 + ')';
         case (d.x > d.parent.x):
-          return 'translate(' + (d.parent.x - d.x + (fontSize * 6) - (fontSize * n)) / (fontSize / 2) + ',' + (d.y - d.parent.y + 5) / 2 + ')';
+          return 'translate(' + (d.parent.x - d.x + (options.linkFontSize * 6) - (options.linkFontSize * n)) / (options.linkFontSize / 2) + ',' + (d.y - d.parent.y + 5) / 2 + ')';
       }
     }
     return 'translate(' + (d.parent.x + 10 - d.x) / 2 + ',' + (d.y - d.parent.y) / 2 + ') rotate(' + angle + ')';
 
 
-  }else {
-    return `translate(${symbsize/4},${symbsize})`
+  } else {
+    return `translate(${options.symbsize / 4},${options.symbsize})`
   }
 }
 
-function rotateFuc(d, o) {
+function rotateFuc(d, o, options) {
   // console.log(d.depth);
   // if (d.depth === 1) { return 'translate(' + o.x(d) + ',' + o.y(d) + ')'; }
   if (d.depth === 0) { return 'translate(' + o.x(d) + ',' + o.y(d) + ')'; }
   //do not rotate if classical styling for fucose
-  var fuctype = document.getElementById('fucoseopt').value;
-  if (fuctype === 'fucoriginal') {
+
+  if (options.orientation === "right-to-left") {
+    return 'translate(' + o.x(d) + ',' + o.y(d) + ')';
+  }
+  if (options.fuctype === 'fucoriginal') {
     return 'translate(' + o.x(d) + ',' + o.y(d) + ')';
   } else if (d.data.monosaccharide !== "Fuc") {
     return 'translate(' + o.x(d) + ',' + o.y(d) + ')';
@@ -419,17 +585,24 @@ function drawsymbol(d, i, type) {
 
 
 //returns the linkage text after replacing the a to alpha and b to beta
-function linkageText(d, abbr) {
+function linkageText(d, i, options) {
   var str = d.data.linkage;
+  if (str == '') { return }
+  if (i===0) {
+    str = str.replace('a', '\u03B1').replace('b', '\u03B2')
+    return str;
+  }
 
-  //special case for fucose to the right
-  if (d.data.monosaccharide === "Fuc" && d.x > d.parent.x) {
+  // console.log(d);
+  let abbr = options.linkAbbr;
+  //special case for fucose to the right this turns the linkage reverse
+  if (options.orientation === "bottom-to-top" && d.data.monosaccharide === "Fuc" && d.x > d.parent.x) {
     if (abbr == true) {
       str = str.replace(/[\d\?]\-/g, '');
     }
     str = str.split("").reverse().join("");
-    str = str.replace (/a/gi, '\u03B1').replace(/b/gi, '\u03B2');
-    console.log(str);
+    str = str.replace(/a/gi, '\u03B1').replace(/b/gi, '\u03B2');
+    //console.log(str);
     return str;
   }
 
@@ -456,7 +629,7 @@ function linkageText(d, abbr) {
       return '\u03B1' + str.charAt(str.length - 1);
     } else if (str.charAt(str.length - 2) === 'b') {
       return '\u03B2' + str.charAt(str.length - 1);
-    }else {
+    } else {
       return str;
     }
   }