+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 All changes to the project will be documented in this file
 
+## 1.1.0 - 2020-7-16
+
+### Fixed
+
+- GlycoCT for [9Ac]Neu5Ac fixed to show acetyl instead of n-acetyl at 9-position.
+- Temporary fix for Issue [#6](https://github.com/akulmehta/GlycoGlyphPublic/issues/6)
+
+### Added
+
+- Thanks to GlyGen team, added links to Glygen for GlyTouCan IDs and from the Glygen API able to retrieve links to pubchem as well as ChEBI for those which are registered with GlyGen.
+- Added ability to rotate the structure and draw from right to left.
+
+### Developer Changes
+
+- Changed `d3glycanstruc` parameters to have an options parameters where you can define multiple things. This will make it easy to override parameters based on application needs for the future.
+
+
 ## 1.0.1 - 2020-03-17
 
 ### Fixed